Replace material-ui with chakra-ui
Material-ui was not playing nice with Next, styles were not persisting
whenever the app had to rerender on the client since class names were
no longer consistent with what was rendered on the server. I spent a
few hours trying to figure out a workaround, gave up, and switched to
chakra after talking with a friend. So far so good.